app/requirements.txt
google-adk==1.10.0
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
googlemaps
